SkiaRenderer/macOS: Add damage and primary plane handling
SkiaOutputDeviceBufferQueue currently holds the assumption that there
always exists a primary plane. It may be that a frame's damage won't
change that plane's contents, but the plane will always be there.

On macOS, this is not the case. Frames will alternate between having
only the primary plane (when not all content can be represented as a
CALayer tree) and having no primary plane (when all content is
represented by the overlay CALayer tree).

Inform SkiaOutputDeviceBufferQueue when there is no primary plane for
the current frame, by making the call to SchedulePrimaryPlane be
unconditional, but with a base::Optional argument. The functions
SwapBuffers, PostSubBuffer, and CommitOverlayPlanes are responsible
for updating SkiaOutputDeviceBufferQueue::current_image_ and
submitted_image_. Add an if-case for those functions when there no
primary plane was scheduled.

Update OverlayProcessorMac to always report full damage to the
SurfaceAggregator. All quads are needed, every frame, to construct
the CALayer tree.

Bug: 894929
Change-Id: Id67f58eeb22a0e92f050f37be900bd6cfa5a7845
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268478
Commit-Queue: ccameron <ccameron@chromium.org>
Reviewed-by: Peng Huang <penghuang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782870}