Avoid setting ui::AXTreeData::parent_routing_id in Blink's serializer
By the time that an AXTreeData structure is received in the browser
process after being sent from a renderer, we should already be able
to retrieve the AXTreeID of its parent AXTreeData just by navigating
the frame tree, that is, by calling GetAXTreeID() over the parent
RFHI of the RFHI that received the message in the browser process.

Thus, we can remove the code in BlinkAXTreeSource setting the parent
frame's routingID along with the code in RFHI that would translate it
into the AXTreeID of the parent AXTreeData, simplifying the code and
helping towards the goal of using the EmbeddingToken as AXTreeID by
removing one more piece in the renderer what would otherwise need to
be migrated as well.

Last, this adds a new private method RFHI::ResolveAXParentTreeID() in
preparation for the future removal of AXContentTreeDataToAXTreeData(),
which is where the |parent_tree_id| was currently being set until now.

Bug: 1094150
Change-Id: I5e1d204e1d30d29c6bde952846195779694aa111
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2254025
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782873}