Register PluginVm apps with AppService but do not show
This changes on the approach taken in crrev.com/c/2251571 where we
stopped notifying app service of changes to be consistent with the
fact that we were not registering pvm apps to start with.

Now we register all pvm apps, but we set them not to show in launcher,
shelf, etc. This will allow us to use AppService to manage icons.

Bug: b/158617133
Bug: b/159050715
Change-Id: I5813a3249af20f81229c6829d9c81e56628bdecb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268897
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782894}