Dedupe LoadingAttrValue with LoadingAttributeValue.
The TokenPreloadScanner::StartTagScanner class has its own
LoadingAttrValue enum, and member of this type.

However, we've introduced a more general LoadingAttributeValue enum,
along with a helper method to convert strings to members of this enum.

The StartTagScanner class should use these more general primitives,
as opposed to providing its own.
This is good so that the preload scanner infrastructure does not
accidentally become out-of-sync with other code that uses the more
general LoadingAttributeValue class.

Bug: 1099123
Change-Id: I7a07e17baa722881b10bd55ae28d10976bfae50b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267518
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Commit-Queue: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782899}