Fix: Fix alignment of toast message in PasswordsListHandler
The CL fixes a typo in a CSS selector used in crrev.com/c/2254281, which
caused a bad alignment of the Undo button.

The CL also merges the selector with #undoLabel, which is actually a
leftover that has been around since M78 (crrev.com/c/1729469).

Bug: 1099236
Change-Id: I8315fafd4e4c20be66324d9cc013872066e9f362
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2266877
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#782915}