Kill renderer for unexpected IPC before commit
In IsInactiveAndDisallowReactivation, we currently log a
DumpWithoutCrashing to monitor crashes from wild.

But, we could receive this from a compromised renderer
sending an IPC from a speculative state. This CL kills the
renderer process in that case when we try to ignore such events

Bug: 1085430
Change-Id: I9b5f482207207ed34e5dc1a3b95c7c762d167d9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2213818
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782948}