[ios] Add codesigning information in generated Xcode project
When using new build system, Xcode requires codesigning information
in the Xcode project to build and run the app (even if they are not
used since the build is done via ninja).

Add support for collecting those information when running "gn gen"
(this requires adding a new exception to exec_script but this cannot
be done at build time since the information is used to generate the
Xcode project when passing --ide=xcode to gen step of gn).

Bug: 1094890
Change-Id: I63dabb836355fb7e2324580d5610872ea4fc33f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267099
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782964}