Disable dns-over-https when an Android device or profile is owned
Call into EnterpriseInfo info to check if the device or profile is owned.
The check may be performed asynchronously and could take some time.
While this check is in progress we'll assume a "not owned" state which
will allow for any existing dns-over-https preferences to be honored.

Testing: Hand tested by installing a device ownership app onto an
emulator and checking Net.DNS.DnsConfig.SecureDnsMode to see if the
value is as expected.

Bug: 1085168
Change-Id: I82690a4c1ff3380fbc3599c71a9fc1fd8a37571c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2239530
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#782999}