Reland "Use wheelTick in forms/resources/common-wheel-event.js"
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. This CL we should replace
eventSender.mouseScrollBy with wheelTick which uses
gpuBenchmarking.smoothScrollByXY in
forms/resources/common-wheel-event.js.

TBR=ellyjones@chromium.org,avi@chromium.org,sky@chromium.org,bokan@chromium.org,lanwei@chromium.org

Bug: 1047176
Change-Id: Ib45dce450fe395d53a443612033305bccef6b19d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268227
Reviewed-by: Lan Wei <lanwei@chromium.org>
Commit-Queue: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783007}