[ios] Allow to use custom build of gn with setup-gn.py
When working in the Xcode project generator part of gn, it is
useful to run setup-gn.py to check that the changes do not
break the Chrome on iOS developers' workflow.

Add a parameter to the script since by default gn comes from
depot_tools as a script that uses //buildtools/mac/gn which
is maintained by cipd (and readonly).

Bug: 619921
Change-Id: Iecd221893966e435cb538463fa48b94a8a6aa774
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270204
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783014}