Remove AXContentTreeData and rely only on AXTreeData now instead
Following up on the steps of crrev.com/c/2254025, we can also remove
the code in BlinkAXTreeSource that sets the |routing_id| field of the
AXContentTreeData that will be sent to the browser process as part
of an AXContentTreeUpdate structure, since we can easily get the
same information in RFHI when the message arrives in the other side.

Since this |routing_id| field was the only part of AXContentTreeData
that was not included in ui::AXTreeData already, this change allows us
to completely drop AXContentTreeData from the codebase along with all
the related .mojom and *_traits.{h,cc} files, which gets us one step
closer to the objective of crbug.com/1094150.

Bug: 1094150
Change-Id: Icc3161ed9e13a8215d99d29cf6d05c66f33418e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260776
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Cr-Commit-Position: refs/heads/master@{#783049}