Added ContentDecryptionModule interface for CdmFactoryDaemon
This is an implementation of media::ContentDecryptionModule that adapts
the calls to a mojo interface for
chromeos::cdm::mojom::ContentDecryptionModule. It also adapts the client
interface chromeos::cdm::mojom::ContentDecryptionModuleClient and
directs those to the callbacks provided when the media::CdmFactory
creates the implementation. This also implements media::Decryptor.

BUG=b:153111783
TEST=Unit tests pass

Change-Id: Ie6b65a7f490d5130548c060fe5b3c96920128e23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2253461
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Xiaohan Wang <xhwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783078}