Defer data hashing in cert provider service
Refactor CertificateProviderService to defer the input data hashing till
the very last point before sending the signature request to the
extension. As a consequence, change the CryptohomeKeyDelegate D-Bus
service to pass the unhashed data. Also kill the call in the
CertStoreInstance Mojo service that was passing pre-hashed data (that
service will anyway be deleted in one of follow-ups - see b/119914122).

This is a preparation step for introducing a new version of the
chrome.certificateProvider API that will pass the unhashed inputs to the
extension.

There's no use-visible changes expected. All changes, except the one of
CertStoreInstance, are backwards-compatible, and the CertStoreInstance
is an unlaunched functionality that will be deleted soon (see
b/119914122).

Bug: 1078761
Change-Id: I656f46c41fdd5fd779e27abe76ee6c01ea2349f2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246177
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Reviewed-by: Polina Bondarenko <pbond@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783089}