Allow QuickAnswersView to gain focus on explicit user actions
This CL allows the QuickAnswersView to gain focus when it is created by
pressing the "Got it" button on the UserConsentView or when an answer is
updated after clicking the Retry button, since these are explicit user
actions.

This allows for a more consistent experience, more so when the
screen-reader is enabled.

Bug: b:152059023
Test: Tested on Atlas test device.
Change-Id: I6a6ba338c2ffbf8b03bdaf4d8adf2f0a158746f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268474
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783095}