wm: Change cross fade animation to use ThroughputTracker.
Removes a test which was a regression test to catch double counting
of animations with two properties (but animations with one property
only single counted which skewed data). It doesn't look like the new
framework will have this problem, and dealing with compositor in unit
tests is tricky.

Test: manual
Bug: 1081986
Change-Id: I2f5624d50c62506f94d1d9a4b90f672349520372
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2268589
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783133}