[base] Explicitly handle our own WM_TIMER msg in ProcessPumpReplacementMessage()
This happens when a native nested loop invokes HandleWorkMessage() =>
ProcessPumpReplacementMessage() which finds the WM_TIMER message
installed by ScheduleNativeTimer(). That message needs to be handle
directly as handing it off to ProcessMessageHelper() causes an
unnecessary BeforeDoInternalWork() which may incorrectly lead the
Delegate's heuristics to conclude that the DoWork() in
HandleTimerMessage() is nested inside a native task.

This is a blocker for
https://chromium-review.googlesource.com/c/chromium/src/+/2226216
where this breaks ThreadControllerWithMessagePumpImpl's heuristic
which sees the previous logic as an application task
nesting in a native task (the automatic nesting detection is
required on the other hand for when a native task truly nests
within another native task -- i.e. can't rely strictly on
ScopedNestableTasksAllower as a hint).

R=fdoray@chromium.org

Bug: 1074019
Change-Id: Ief078e55fd84555e2d194f67168bf51c9afcaa63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270708
Commit-Queue: Gabriel Charette <gab@chromium.org>
Commit-Queue: François Doray <fdoray@chromium.org>
Auto-Submit: Gabriel Charette <gab@chromium.org>
Reviewed-by: François Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783166}