Trust Tokens: Attach Blink-provided additional signing data to requests
We're extending the prototype Trust Token API to allow binding
additional client-provided data to signed outgoing requests. This is
useful, for instance, when you want to sign outgoing requests over the
initiating URL stripped in some client-specified manner. In order to add
this, we expand the interface, adding an additionalSigningData parameter
that takes a string. This passes its data to the network stack, where
it's packed into a new outgoing browser-provided Trust Tokens-specific
header; clients who wish to add specify additional signing data can then
do so by simultaneously (1) providing a value to the
additionalSigningData argument and (2) adding
Sec-Trust-Tokens-Additional-Signing-Data to the list of outgoing headers
to sign.

This CL, the second in a series of two, updates the network service to
take additional signing data values passed from Blink and attach them to
outgoing requests in the new header, implicitly adding the header name
to the list of headers to sign in this case. It also contains a small
refactor adding a minimal collection of public Trust Tokens constants to
the network service's public/cpp folder. (Some of the remaining
constants in the internal constants file are arguably part of the
network service API too and can be moved at a later date once it's
necessary to access them.)

R=csharrison

Bug: 1094012
Change-Id: I6279151da18d3de30994bd717d74472d4370e654
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265518
Commit-Queue: David Van Cleve <davidvc@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Asanka Herath <asanka@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783172}