SharedImageBackingGLImage: Bind texture lazily
There are advantages to not binding a GLImage-backed SharedImage
to a GL texture immediately. In the case of Metal or software
raster, we will not actually need a GL texture.

Lazy-binding GLImages to textures was prevented by a bug whereby
other binding paths exposed through the command buffer would not
update the GLImage's internal binding state, resulting in
DCHECKs.

Fix the command buffer bug, and remove the early bind from
SharedImageBackingGLImage.

Bug: 1092155
Change-Id: I201f901722b93e6a73e866bf8249130148c5673e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2255312
Reviewed-by: Geoff Lang <geofflang@chromium.org>
Commit-Queue: ccameron <ccameron@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783187}