Eliminate hardcoded SkColorSet* use from TabStrip
This is part of the color pipeline work. This is to make sure that
all colors would go through standard color getter paths and can
update correctly when the underlying theme changes.

This change adds a new border color ID used in TabStrip::OnPaintBorder.
Currently the color is the same.

Bug: 1056950
Change-Id: I90ed62bb83c665ecba81eb3840fd8e0798c2ae84
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2265373
Commit-Queue: Maria Villarreal <mavill@microsoft.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Ahmed Mehfooz <amehfooz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783246}