Eliminate hardcoded SkColorSet* use from MdTextButton
This is part of the color pipeline work. This is to make sure that
all colors would go through standard color getter paths and can
update correctly when the underlying theme changes.

This change adds a new disabled border color ID used as the stroke
color in MdTextButton::UpdateColors.
The color will be slightly different (a little bit more opaque).
The previous values are #F5F6F7 on light theme and #2D3032 on
dark theme. With this change, the new colors are #EAEAEA on light
theme and #363739 on dark theme.

Bug: 1056950
Change-Id: I9574ec6c566db2526c42161ea59e92e7fea46da4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261873
Commit-Queue: Maria Villarreal <mavill@microsoft.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783253}