[Nearby] Add task scheduler classes
The scheduler handles scheduling recurring tasks, scheduling retries for
failed requests, delaying requests until network connectivity is
established, queuing up pending requests if scheduling has not been
explicitly started yet or if another request in currently in progress,
and persisting scheduling data to prefs.

When a task is ready to be run, the scheduler notifies the manager
class via a callback to actually perform the task. The scheduler then
waits for the manager to return the result--success or failure--of the
task.

Fixed: b/154864199
Change-Id: Ic9ac981c5b4462641aa789c20a3ca78ff0f13aeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2214614
Commit-Queue: Josh Nohle <nohle@chromium.org>
Reviewed-by: James Vecore <vecore@google.com>
Cr-Commit-Position: refs/heads/master@{#783262}