Support SHELL tests in chromium
SHELL tests weren't being supported in chromium due to some
incompatibility with C++ version. However, it was safe also
as there's another checkout in google3 which was working as
expected. Also, some of the fixes have been uploaded to
google3 not to the upstream yet.

BUG=chromium:1095129

Change-Id: I44fe7f575d5bc67210bfad3ace2ddd0aa5bed3da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2246145
Reviewed-by: Nico Weber <thakis@chromium.org>
Commit-Queue: Amr Aboelkher <amraboelkher@google.com>
Cr-Commit-Position: refs/heads/master@{#783265}