[FragmentItem] Avoid crash in ComputeSelfInkOverflow
This patch avoids crash in |ComputeSelfInkOverflow| when
corresponding |LayoutObject| was destroyed or moved.

Unlike <crbug.com/1095577>, which was about stale block
children and was fixed by using |PostLayoutChildren()| in
r779382 <crrev.com/c/2249281>, this crash is about an
atomic inline was moved or destroyed, but we still have a
fragment for it.

I don't have good explanation about this, except that the
layout tree is dirty in the pre-paint phase.

This patch is speculative.

Bug: 1099613
Change-Id: I4f9eb6714e42ff9521e456d9a35e0726de604a83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269629
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783332}