Fix StartServiceWorker crash for installing service worker.
Currently if you call ServiceWorkerContextWrapper::StartServiceWorker(..)
while a SW for the given scope is installing the browser will crash because
registration->active_version() == nullptr.
Remove my fix and you'll see the crash in the test I added.
Since StartServiceWorker is a public method of a relatively high level class
it is supposed to be safe to use it under any reasonable circumstances.

What can be done to fix it?
a) Return an error from StartServiceWorker. This is what I'm doing here.
b) Delay the start till the installation gets finished. This might be harder to implement

R=bashi, falken

Change-Id: I558dc4d12d0a91ea33451563ff70550986b957e8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260753
Commit-Queue: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783386}