DPWA: Early exit Shortcuts Menu registration when WebApp update fails
This CL adds a check in
WebAppInstallTask::OnUpdateFinalizedRegisterShortcutsMenu() for the
InstallResultCode from updating the WebApp. If the app update fails, we
early exit and avoid running Shortcuts Menu Registration or
Unregistration code.

This should fix crashes reported in the RegisterShortcutsMenuWithOs code
running as part of app updates. See bug below.

The crashes were caused by attempting to BuildShortcutInfo() with an
app_id corresponding to an app that is not in the WebAppRegistrar.

Bug: 1097743
Change-Id: Iecd62e424f8be3652bbb1bd3a2d07f611fdb804f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2271242
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Reviewed-by: Alexey Baskakov <loyso@chromium.org>
Commit-Queue: Rahul Singh <rahsin@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#783395}