Use only 'keydown' event to close Android IME.
Before this CL, both of 'keydown' and 'keyup' events are marked as
consumed by ArcInputMethodManagerService when the back button on the
shelf is pressed while Android IME is shown.
It triggers never-ending key repeat because 'keydown' events are sent to
post IME stage, but 'keyup' events are not.

Bug: 1099673
Change-Id: I14c2496cbc0cb7b49641bd4642647fb7388d765a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2269626
Commit-Queue: Yuichiro Hanada <yhanada@chromium.org>
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783405}