Simplify/improve static data handling in LocalFrameUkmAggregator
We can use plain string literals here rather than creating String
objects (which has a cost both from a runtime and footprint PoV). This
also makes it trivial to convert the MetricInitializationData vector to
a static array, which makes it easy to add additional compile-time
checking that it is the right size. We can wrap the array in a
base::span to facilitate iteration.
Similarly convert the threshold vector to an array.

Change-Id: I09962fe95f0f76e05c3e5a3797aa97ea2c78e566
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270299
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Stephen Chenney <schenney@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783450}