Fix event listener check in cert provider service
Fix discrepancy between the event name that
CertificateProviderServiceFactory subscribes to and the event name that
is later checked for listeners presence. This fixes the logical error
that we're making the decision based on the presence/absence of the
event listener that we're not listening to.

This is assumed to be a low-priority bug, since the commonly used
real-world extensions never unsubscribe from the certificateProvider
events and therefore never trigger the buggy scenario. Also the main
consequence would be the loss of certificates from that extension, which
seems OK in this half-shutdown state.

Bug: 1067683
Change-Id: I9bb8486ebb6f83af0bb845f6652a5d2d658d779e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2247019
Reviewed-by: Fabian Sommer <fabiansommer@chromium.org>
Commit-Queue: Maksim Ivanov <emaxx@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783549}