CrOS: memory: improve userfaultfd API
This CL improves the userfaultfd wrapper API to include the number of
bytes copied/zeroed. This is necessary because there may be situations
where we cannot either copy because there is a pending remap, unmap, or
remove so the copy/zero will always fail with EAGAIN until that event is
handled.

Additionally, there is a situation where the ioctl for
UFFDIO_COPY/ZEROPAGE can return -EAGAIN even though it actually
completed the copy and this can happen when it cannot wake the range.
This change makes the API more correct.

BUG=1067833

Change-Id: Ie4a8d7a96b932ce2c025221961744e2fd926ef6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270624
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783554}