Introduce settings-avatar-icon
It seems that there's currently some code duplication in the WebUI
codebase to generate round avatars of the signed-in user [1], so this
CL adds a component to do this. For the moment it lives under
autofill_page/ and is only used for passwords settings pages. Future CLs
will move it to people_page/, and remove the duplication from other
sites.

[1] https://source.chromium.org/search?q=%5C.iconUrl%20f:settings%20-f:%5Eout%7C%5Echrome%2Ftest

Bug: 1049141
Change-Id: Id6149bd6be9e3ab1183f0f53778ed3e472f76e15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270319
Auto-Submit: Victor Vianna <victorvianna@google.com>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783590}