[ProfilePicker]: Add navigation mechanism to the profile picker.
This CL does the following:
(1) Adds a navigation mechanism for the different screens of the
picker:
- Main view to pick, edit, add profile.
- Profile creation flow: profile type choice, local customization,
signin. The view to be shown in the first step of
'chrome://profile-picker/new-profile' is not fixed and will depend on a
number of factors.

(2) Creates one of the profile creation views for the new profile type
choice (Signed In or Local). The profile creation flow should be lazy
loaded since the picker is shown on startup and the creation flow is
only needed if the user decides to create a new space/profile.

Bug: 1063856
Change-Id: Ic59fb34ff6a9fc8fbfcb087c22cdfb3f8c9e689b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2260616
Commit-Queue: Monica Basta <msalama@chromium.org>
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783601}