Refactor FeedbackCollector.
For feedback on the Discover feed (on the NTP), we want to drop all
the PSD other than the ones directly relevant for the feed.  The
feedback will be processed by the discover team instead of the
chrome team, and we want to make sure PII is not leaked to teams
outside chrome to strongly comply with privacy.

The approach we took was to make FeedbackCollector a base class.
For feedback going to chrome we use a derived class called
ChromeFeedbackCollector.  For feedback going to the AGSA team,
we use a derived class called FeedFeedbackCollector, which only
includes PSD specific to the Chrome feed.

A new unit test for the FeedFeedbackCollector has been added.

We also removed the FeedContext parameter from the FeedbackCollector,
since it is only used now by the FeedFeedbackCollector.

(cherry picked from commit ba6b88573d1faebda31c30ecf71249401a2a7e05)

Bug: 1091058
Change-Id: Ia088b20aaf2ef4a50c2ce163cacd159f7a0c28e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2241455
Commit-Queue: Peter Williamson <petewil@chromium.org>
Reviewed-by: Tommy Nyquist <nyquist@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#782810}
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2273759
Reviewed-by: Peter Williamson <petewil@chromium.org>
Cr-Commit-Position: refs/branch-heads/4183@{#26}
Cr-Branched-From: 740e9e8a40505392ba5c8e022a8024b3d018ca65-refs/heads/master@{#782793}