Omnibox UI: animate URL elisions from both ends of URL
This CL adds a field trial option to elide not just the path, but also
subdomains beyond the registrable domain. To support this option, the
animation has to get fancier: we can now longer fade the color of the
path, but must instead animate the display rect and offset. (The mocks
also have a fade that happens while the display rect is changing, but
I haven't implemented that yet.)

I've replaced PathFadeAnimation with ElideAnimation that elides or
unelides to a particular range of the string. Because the animation
sets and can query state on the RenderText (the display rect and
display offset), we can collapse the hover fade-in and fade-out
animations into one animation.

I've also cleaned up a couple other things along the way, mostly
removing the extra has_started_ state on the animation and replacing
it with GetCurrentValue().

Bug: 1090393
Change-Id: Iccd3ba52b74152bf207c440a549a8314da3da704
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2261893
Commit-Queue: Emily Stark <estark@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783666}