Address an ambiguity in @WrappedPath() in GN generate_wrapper.
In the GN generate_wrapper() template, it turns out there is
an ambiguity with the @WrappedPath() syntax: if you pass in
something (say, @WrappedPath(./foo)) that ends up resolving to
the current directory, it's unclear if the resolved path should be
'./foo' or 'foo'. The prior code would resolve it to 'foo', but
that causes a problem for scripts like test_env.py that want to
invoke the path as an executable; passing the result to something
like os.Popen() will cause the interpreter to search $PATH for
'foo', rather than invoking './foo' directly, and if '.' isn't in
PATH, you get an error as a result.

This CL changes the logic to return './foo' instead; hopefully
there are no cases where the value needs to be 'foo'. If there are;
we'll need to adopt a different approach.

Change-Id: I4905eba1280ed39e79140a6a850ef8139279dfdd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2264931
Commit-Queue: Dirk Pranke <dpranke@google.com>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Reviewed-by: Andrew Grieve <agrieve@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783670}