inclusion: rename "master_preferences" &c, part 1/n
This change renames:
* master_prefs::MasterPrefsPath -> initial_prefs::InitialPrefsPath
* first_run::internal::MasterPrefsPath -> ...::InitialPrefsPath

This is the first part of renaming the "master preferences" file and all
the code referencing it. A file in a platform-dependent location named
"master_preferences" or similar has historically been the documented way
to impose initial preferences in newly-created profiles. Removing
support for the old name would break backward compatibility, so this
sequence of CLs will instead:

1) Rename the concept from "master preferences" to "initial preferences"
   throughout the codebase *without* changing the name of the actual
   initial preferences file on disk
2) Add support for a secondary initial preferences file, named
   "initial_preferences"
3) Update all public documentation to refer to the new
   "initial_preferences" file instead

It is not likely that it will ever be possible to drop support for the
old filename, given its likely wide use, but this will at least improve
things.

Bug: 1097204
Change-Id: Ic4b3ed62c32171c505b287171fb4213a0950a91f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267422
Commit-Queue: Avi Drissman <avi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Gabriel Charette <gab@chromium.org>
Reviewed-by: Greg Thompson <grt@chromium.org>
Reviewed-by: Robert Sesek <rsesek@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783671}