[Hotseat] Using layer target bounds for comparison
When SetTranslucentBackground() is called, hotseat background may have
the running bounds animation. Then it is not meaningful to compare
|background_bounds|, the parameter of SetTranslucentBackground(), with
the layer's current bounds. The layer's target bounds should be compared
with instead.

In this particular bug, AppWindow::Init() adds the app icon to shelf
via AppWindowRegistry::AddAppWindow() then may remove the app icon from
shelf through AppWindow::Hide(). Due to the problem showed above, when
the shelf item is removed, the hotseat bounds animation triggered by
shelf item addition is still running so the hotseat background's bounds
do not update yet. As a result, the target bounds for shelf item removal
are identical with the current layer bounds. So it fails to update
bounds.

Bug: 1097645
Change-Id: I4046792060797faedb9f358ba3e4706338e5b7f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274144
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783714}