Update FrameTokenMessageQueue FrameToken Checks
FrameTokenMessageQueue was used as a way to diagnose incorrect frame
submission ordering bug within the Renderer and Viz. The root cause
behind that (crbug.com/1051625) was fixed. We never got around to
removing the DumpWithoutCrashing.

These have popped back up as side-effects of other crashes, but were not
needed to identify nor debug those.

There is also a race condition which can lead to this being a false
signal. Where FrameTokenMessageQueue::Reset can be called before Viz has
submitted an ACK. A proper fix for this would be to change from raw
ints, to an expansion of LocalSurfaceId. So that the source of the frame can be
attributed.

However that's a larger effort. Instead this change removes some of the
previous debugging, as it is no longer needed. While also updating the
check in DidProcessFrame to only be ran when the are queued items, and
it isn't the ack from the Reset. This should hopefully reduce the noise.

TEST=FrameTokenMessageQueueTest

Bug: 1058439
Change-Id: I6854b1e5e71cdc77c72f898adc81541047592c9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270837
Commit-Queue: Jonathan Ross <jonross@chromium.org>
Reviewed-by: kylechar <kylechar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783720}