Reland "Use wheelTick in forms/resources/common-wheel-event.js" Part 2
After we finish the scroll unification, the scrolls happens mainly on
the compositor thread, and the scroll code in the main thread will be
removed. eventSender sends the scroll events to main thread, so it
would not work after the scroll unification. This CL we should replace
eventSender.mouseScrollBy with wheelTick which uses
gpuBenchmarking.smoothScrollByXY in
forms/resources/common-wheel-event.js.

Because the original CL
https://chromium-review.googlesource.com/c/chromium/src/+/2171955 has
changes in synthetic_gesture_target_mac.mm which will send the wheel
events to the right window, for example, if the mouse is over the popup
window, the events should be sent to popup window. Also, it rewrites some
layout tests using gpuBenchmarking.smoothScrollByXY. Some of the tests
are flaky on Win 7 because of the animation does not finishes before the
tests finish. This CL only has the change to all the related the layout
tests and also add them to the test expectation in case they are still
flaky.

TBR=bokan@chromium.org

Bug: 1047176
Change-Id: I80600a9511f7a3094bc5ff16f9c2fa048f5150d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2270850
Commit-Queue: Lan Wei <lanwei@chromium.org>
Reviewed-by: Lan Wei <lanwei@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783736}