Allow menu-item selections to be unselected
When a Quick-Answer widget is showing above a context-menu within the
browser, pressing the Up key will transfer the focus to this widget if
the first menu-item is selected at the time.

While it visually looks like the selected-item was cleared, pressing the
Down key again selects the second menu item instead of selecting the
first item again.

This CL fixes that by adding an API to select the parent menu-item to
update the menu's pending state as well, instead of only unselecting the
current menu-item which would not affect the menu's state.

Bug: 152057976
Test: Tested on Chrome OS VM.
Change-Id: Id73682254a9503c0a458e7525204138859f6bb1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2267454
Commit-Queue: Abhijeet Singh <siabhijeet@google.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#783801}