crostini: Fix 2/3 of the the Crostini settings test flakes
Sometimes (when the system is under sufficient load, tests are running
slow enough, the system is tickled the right way by swarming, and other
things I haven't figured out yet), there is a chance that the Crostini
Settings Subpage will be attached to the DOM (or at least, the attached
method gets fired) *before* the SubPageDetails suite runs its setup to
navigate to the page. When this happens the request for crostini
installer status (fired by the subpage details attached event) happens
before we attach a listener in the suite setup, so our listener hangs
forever for a message that will never arrive.
That behaviour isn't actually required for the tests to run so move the
check from the suite setup into its own test, and check in a way that if
it fails it fails, instead of hanging forever.

Bug: chromium:1082575
Test: Browser tests
Change-Id: I590a0aab8b6e4e0bfba2497a1706667201dfa3b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2274404
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Reviewed-by: Daniel Ng <danielng@google.com>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Commit-Queue: David Munro <davidmunro@google.com>
Cr-Commit-Position: refs/heads/master@{#783805}